Include the possibility to add HTML on the cells.
For backward compatibility, by default, it still uses escapehtml. A new
property was added to the Table model, to keep which plugin must be used
to escape the content of the table.

When it is set to null `$table->setEscape(null)`, it doesn't apply any
change/parsing, thus allowing the insertion of HTML into  the cells.